+++ projects/ngx-material/src/components/index.ts
@@ -1,3 +1,5 @@
 
 export * from './loading-overlay/loading-overlay.component';
 export * from './header/header.component';
+export * from './divider/ngx-divider.component';
+export * from './icon-button/icon-button.component';