+++ projects/ngx-material/src/stories/buttons/button-preview/button-preview.component.ts
@@ -1,0 +1,17 @@
+import {ChangeDetectionStrategy, Component} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {IconButtonComponent} from "../../../components/icon-button/icon-button.component";
+
+@Component({
+  selector: 'ngx-button-preview',
+  standalone: true,
+  imports: [CommonModule, IconButtonComponent],
+  templateUrl: './button-preview.component.html',
+  styleUrls: ['./button-preview.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class ButtonPreviewComponent {
+
+  active = false;
+
+}